+++ sources/packages/backend/apps/api/src/services/education-program-offering/education-program-offering-validation.models.ts
@@ -33,6 +33,7 @@ import {
   PeriodMaxLength,
   IsDateAfter,
   ValidationResult,
+  ValidationContext,
 } from "../../utilities/class-validation";
 import { Type } from "class-transformer";
 import { ProgramAllowsOfferingIntensity } from "./custom-validators/program-allows-offering-intensity";
@@ -197,63 +198,6 @@ export enum OfferingValidationWarnings {
 export enum OfferingValidationInfos {
   InvalidStudyBreakAmountOfDays = "invalidStudyBreakAmountOfDays",
   InvalidStudyBreaksCombinedThresholdPercentage = "invalidStudyBreaksCombinedThresholdPercentage",
-}
-
-/**
- * Represent the context of an error that has an additional context to
- * express a possible warning or some relevant information captured during
- * the validation process.
- * All validations when failed will generate an error. The ones that have
- * the warning or information contexts will not be considered critical.
- */
-export class ValidationContext {
-  /**
-   * Creates an error context that will make the error downgrade to
-   * a condition of a warning information will force the offering
-   * to be reviewed by the Ministry.
-   * @param warningTypeCode warning code that uniquely identifies this condition.
-   * @returns the warning context.
-   */
-  static CreateWarning(
-    warningTypeCode: OfferingValidationWarnings,
-  ): ValidationContext {
-    const newContext = new ValidationContext();
-    newContext.type = ValidationContextTypes.Warning;
-    newContext.typeCode = warningTypeCode;
-    return newContext;
-  }
-
-  /**
-   * Creates an error context that will make the error downgraded to
-   * a condition of a simple information that will still allow the
-   * offering to be automatically approved.
-   * @param infoTypeCode information code that uniquely identifies this condition.
-   * @returns the information context.
-   */
-  static CreateInfo(infoTypeCode: OfferingValidationInfos): ValidationContext {
-    const newContext = new ValidationContext();
-    newContext.type = ValidationContextTypes.Information;
-    newContext.typeCode = infoTypeCode;
-    return newContext;
-  }
-
-  /**
-   * Context type.
-   */
-  type: ValidationContextTypes;
-  /**
-   * Unique identifier of the validation error.
-   */
-  typeCode: OfferingValidationWarnings | OfferingValidationInfos;
-}
-
-/**
- * Types of non-critical errors that will not prevent
- * the offering from being saved.
- */
-export enum ValidationContextTypes {
-  Warning = "warning",
-  Information = "information",
 }
 
 /**
